Role Controller Authorization Complete

+++ app/Http/Controllers/Backend/RoleController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
 use App\Http\Controllers\Controller;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Permission;
 
 class RoleController extends Controller
@@ -17,8 +18,13 @@ class RoleController extends Controller
      */
     public function index()
     {
-        $roles = Role::all();
-        return view('backend.roles.index', compact('roles'));
+        if (Auth::user()->can('role-view')) {
+            $roles = Role::all();
+            return view('backend.roles.index', compact('roles'));
+        } else {
+            return redirect()->route('admin.401');
+        }
+        
     }
 
     /**
@@ -28,8 +34,13 @@ class RoleController extends Controller
      */
     public function create()
     {
-        $all_permissions = Permission::all();
-        return view('backend.roles.create', compact('all_permissions'));
+        if (Auth::user()->can('role-create')) {
+            $all_permissions = Permission::all();
+            return view('backend.roles.create', compact('all_permissions'));
+        } else {
+            return redirect()->route('admin.401');
+        }
+        
     }
 
     /**
@@ -40,13 +51,17 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        $role = Role::create(['name'=> $request->name]);
-        $permissions = $request->input('permissions');
-        if (!empty($permissions)) {
-            $role->syncPermissions($permissions);
+        if (Auth::user()->can('role-create')) {
+            $role = Role::create(['name'=> $request->name]);
+            $permissions = $request->input('permissions');
+            if (!empty($permissions)) {
+                $role->syncPermissions($permissions);
+            }
+            Toastr::success('Role Created Successfully', 'Success');
+            return redirect()->route('admin.role.index');
+        } else {
+            return redirect()->route('admin.401');
         }
-        Toastr::success('Role Created Successfully', 'Success');
-        return redirect()->route('admin.role.index');
 
     }
 
@@ -69,9 +84,13 @@ class RoleController extends Controller
      */
     public function edit($id)
     {
-        $role = Role::findById($id);
-        $all_permissions = Permission::all();
-        return view('backend.roles.edit', compact('role', 'all_permissions'));
+        if (Auth::user()->can('role-create')) {
+            $role = Role::findById($id);
+            $all_permissions = Permission::all();
+            return view('backend.roles.edit', compact('role', 'all_permissions'));
+        } else {
+            return redirect()->route('admin.401');
+        }
     }
 
     /**
@@ -83,16 +102,20 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $role = Role::findById($id);
-        $permissions = $request->input('permissions');
+        if (Auth::user()->can('role-update')) {
+            $role = Role::findById($id);
+            $permissions = $request->input('permissions');
 
-        if (!empty($permissions)) {
-            $role->name = $request->name;
-            $role->save();
-            $role->syncPermissions($permissions);
+            if (!empty($permissions)) {
+                $role->name = $request->name;
+                $role->save();
+                $role->syncPermissions($permissions);
+            }
+            Toastr::info('Role Updated Successfully', 'Info');
+            return redirect()->route('admin.role.index');
+        } else {
+            return redirect()->route('admin.401');
         }
-        Toastr::info('Role Updated Successfully', 'Info');
-        return redirect()->route('admin.role.index');
     }
 
     /**
@@ -103,9 +126,13 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
-        $role = Role::findById($id);
-        $role->delete();
-        Toastr::warning('Role Deleted Successfully', 'Danger');
-        return back();
+        if (Auth::user()->can('role-delete')) {
+            $role = Role::findById($id);
+            $role->delete();
+            Toastr::warning('Role Deleted Successfully', 'Danger');
+            return back();
+        } else {
+            return redirect()->route('admin.401');
+        }
     }
 }
